Version 5.1.0,  Revision 20375,  Hash 88616e0bc
MODIFIED *  include/driver/version.h sbin/compilation/check_updated_locks.sh sbin/compilation/check_updated_sources.sh sbin/compilation/object_save_restore_remove.sh

RENAMED *  sbin/compilation/check_object.sh -> sbin/compilation/check_object_childs.sh

Bugs:
- Fixed SAVE&REPORT. Some source are intrinsically needed to be recompiled. A better RESTORE folder name would solve this problem.

Patch sent by:  Andrea Marini <andrea.marini@cnr.it>

+++ sbin/compilation/object_save_restore_remove.sh
@@ -36,8 +36,10 @@ fi
 #
 ldir=`dirname $full_path`
 llib=`basename $ldir`
+#
 source ./sbin/compilation/stamp_remove.sh "lib"
-if [ "$operate" == "remove" ]  ; then
+#
+if [[ "$operate" == "remove" ]] && [[ -f $full_path ]]  ; then
   if [ "$VERB" == 1 ] ; then
    echo "rm -f  $full_path"
    echo "rm -f config/stamps_and_lists/lib${llib}.a.stamp"
@@ -46,23 +48,19 @@ if [ "$operate" == "remove" ]  ; then
    rm -f config/stamps_and_lists/lib${llib}.a.stamp 
   fi
 fi
-if [ "$operate" == "save" ] ; then
-  if [ -f $full_path ] ; then
-    if [ "$VERB" == 1 ] ; then
-     echo "mkdir -p $ldir/$save_dir"
-     echo "mv $full_path $ldir/$save_dir"
-    else
-     mkdir -p $ldir/$save_dir
-     mv $full_path $ldir/$save_dir
-    fi
+if [[ "$operate" == "save" ]] && [[ -f $full_path ]]  ; then
+  if [ "$VERB" == 1 ] ; then
+   echo "mkdir -p $ldir/$save_dir"
+   echo "mv $full_path $ldir/$save_dir"
+  else
+   mkdir -p $ldir/$save_dir
+   mv $full_path $ldir/$save_dir
   fi
 fi
-if [ "$operate" == "restore" ] ; then
-  if test -d "$ldir/$restore_dir"; then
-    if [ "$VERB" == 1 ] ; then
-     echo " mv $ldir/$restore_dir/$file $ldir"
-    else
-     mv $ldir/$restore_dir/$file $ldir
-    fi
+if [[ "$operate" == "restore" ]] && [[ -f $ldir/$restore_dir/$file ]]  ; then
+  if [ "$VERB" == 1 ] ; then
+   echo " cp $ldir/$restore_dir/$file $ldir"
+  else
+   cp $ldir/$restore_dir/$file $ldir
   fi
 fi
